refactor: Add Switch component for UI toggle functionality
Add a new Switch component to the UI for toggle functionality. This component will be used in the LinkEditorPanel to enable the "open in new tab" functionality. The Switch component is imported from the "@radix-ui/react-switch" package and is styled with CSS classes for different states. This refactor improves the readability and maintainability of the codebase.